Update settings and manual for code formatting and venv
More recent VS Code versions have much better built-in support
for linting and code formatting of Python code. This change takes
this into consideration and configures Black for use.

Additionally creation of virtual environments is integrated in the
Python extension and doesn't need any dedicated task. This is removed
from the tasks.json